gps: Standby support
In standby mode, the gps consumes less power (1mA vs ~25mA in tracking mode). This can help prolong battery life if support for backup mode is not implemented. To get GPS to enter standby, send a downlink of 0x00 and to exit standby, send a downlink of 0x01. This will be changed in the future to enter standby after a transmission so the entire board can enter sleep.