Rework BMP280
The previous library didn't work well when trying to deepsleep. Upon sleeping, the state of I2C_1 and I2C_2 of the STM32WLExx is cleared. This means that we can't sleep while they are in use. We will create a new I2C class whenever we want to communicate with the bmp280 and free it when we are done. This means we can sleep in between.
NOTE: Some changes have been made to mbed os itself as well. See https://github.com/hallard/mbed-os/commit/fabbed360899e1ae10362da9c28333c315930355
See https://github.com/ARMmbed/mbed-os/issues/15191